+++ CHANGES.rst
@@ -8,8 +8,13 @@ Changelog
 * Add documentation for methods :meth:`RecordList.exists` and
   :meth:`RecordList.ensure_one`.
 
+* Add documentation for methods :meth:`RecordList.mapped`,
+  :meth:`RecordList.filtered` and :meth:`RecordList.sorted`.
+
 * Changed method ``exists`` on :class:`RecordList` and :class:`Record`
   to return record(s) instead of ids.
+
+* Fix methods ``mapped``, ``filtered`` and ``sorted``. Add tests.
 
 * Fix method ``RecordList.ensure_one()`` when there's identical ids
   or ``False`` values.